requirements.txt
jujuclient
pyyaml
